Fixed invalid package names
